ENCORE PLUS DE DOUBLE POINTEUR

+++ Module 4/Code/donnees.h
@@ -1,6 +1,6 @@
 struct infos {
     int pouls;
-    int temps;
+    long temps;
 };
 
 //typedef struct infos infos;
